// app/[locale]/customer-dashboard/types/subscription.ts
export interface Place {
    id: string
    name: string
    location: string
    phoneNumber: string
  }
  
  export interface Subscription {
    id: string
    name: string
    benefits: string
    price: number
    website?: string | null
    visitsPerMonth?: number | null
    places: Place[]
  }